services/ml_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[ext]
asyncpg
pydantic
pydantic-settings
python-dotenv
aiokafka
psycopg2-binary
torch
numpy
scikit-learn
joblib
httpx
pytest-asyncio
